Refactor UserPageComponent: Implement pagination and modal components
- Removed inline pagination and modal logic from UserPageComponent.
- Created PaginationUserComponent to handle pagination logic and UI.
- Created ModalEditUserComponent for user editing functionality.
- Updated getAllStudents action to change default limit from 5 to 7.
- Added new TypeScript interfaces for ModalEditUserComponentProps and PaginationUserComponentProps.
- Adjusted UserPageComponent to utilize new components and improved loading state handling.
